requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.0.0
requests==2.21.0
soupsieve==1.9.1
SQLAlchemy==1.3.3
urllib3==1.24.2
Werkzeug==0.15.2
